+++ NewStructure/CMakeLists.txt
@@ -214,6 +214,15 @@ else()
   message(WARNING "HGCROC calib sample parser will not be built - h2g_decode library not found")
 endif()
 
+#Evaluate reco effi HGCROC
+if(H2G_DECODE_FOUND)
+  add_executable(EvaluateRecoEffi EvaluateRecoEffi.cc EvaluateRecoEffiHGCROC.cc)
+  target_compile_definitions(EvaluateRecoEffi PRIVATE DECODE_HGCROC)
+  target_link_libraries(EvaluateRecoEffi LFHCAL Analysis WaveformFitting ${ROOT_LIBRARIES} h2g_decode)
+else()
+  message(WARNING "HGCROC reco effi evaluation will not be built - h2g_decode library not found")
+endif()
+
 
 # Set default build target
 set_property(DIRECTORY PROPERTY VS_STARTUP_PROJECT DataPrep)
